refactor: useAddressNames
