+++ packages/maskbook/src/social-network-adaptor/twitter.com/injection/useEthereumName.ts
@@ -1,4 +1,5 @@
-import { useResolveENS, useResolveUNS } from '@masknet/web3-shared'
+import { first } from 'lodash-es'
+import { useResolveENS, useResolveUNS, useAddressNames, useENSLabel } from '@masknet/web3-shared'
 import { useEffect, useMemo, useState } from 'react'
 import type { AsyncState } from 'react-use/lib/useAsyncFn'
 
@@ -32,6 +33,9 @@ export function useEthereumAddress(nickname: string, twitterId: string, bio: str
     const name = useEthereumName(nickname, twitterId, bio)
     const { value: addressENS, loading: loadingAddressENS } = useResolveENS(name)
     const { value: addressUNS, loading: loadingAddressUNS } = useResolveUNS(name)
+    const { value: names = [], loading: loadingAddressNames } = useAddressNames(twitterId)
+
+    const ownerAddress = first(names)?.ownerAddress ?? ''
 
     useEffect(() => {
         setAddress('')
@@ -39,7 +43,7 @@ export function useEthereumAddress(nickname: string, twitterId: string, bio: str
         if (matched) setAddress(matched[0])
     }, [bio])
 
-    const isLoading = loadingAddressENS || loadingAddressUNS
+    const isLoading = loadingAddressENS || loadingAddressUNS || loadingAddressNames
 
     return {
         loading: isLoading,
@@ -47,8 +51,8 @@ export function useEthereumAddress(nickname: string, twitterId: string, bio: str
             ? undefined
             : {
                   type: addressENS ? 'ENS' : addressUNS ? 'UNS' : 'address',
-                  name,
-                  address: isLoading ? '' : addressENS ?? addressUNS ?? address ?? '',
+                  name: addressENS || addressUNS ? name : '',
+                  address: isLoading ? '' : addressENS ?? addressUNS ?? ownerAddress ?? address ?? '',
               },
         error: undefined,
     } as AsyncState<{
