+++ packages/web3-shared/src/hooks/index.ts
@@ -10,7 +10,7 @@ export * from './useNonce'
 export * from './useGasPrice'
 export * from './useEtherPrice'
 export * from './useENSLabel'
-export * from './useENSDomains'
+export * from './useAddressNames'
 export * from './useERC1155TokenAssetDetailed'
 export * from './useERC165'
 export * from './useERC20TokenAllowance'
